splitting of tabs from Mainwindow file
