+++ gui/__init__.py
@@ -4,5 +4,16 @@ from .gui import MainWindow
 from .Ui_MainWindow import Ui_MainWindow
 from .effect_selection import EffectSelectionDialog
 from .StageView import StageView
+from .tabs import BaseTab, ConfigurationTab, FixturesTab, ShowsTab, StageTab
 
-__all__ = ['MainWindow', 'Ui_MainWindow', 'EffectSelectionDialog', 'StageView']
+__all__ = [
+    'MainWindow',
+    'Ui_MainWindow',
+    'EffectSelectionDialog',
+    'StageView',
+    'BaseTab',
+    'ConfigurationTab',
+    'FixturesTab',
+    'ShowsTab',
+    'StageTab'
+]
