+++ gui/tabs/__init__.py
@@ -1,0 +1,9 @@
+# gui/tabs/__init__.py
+
+from .base_tab import BaseTab
+from .configuration_tab import ConfigurationTab
+from .fixtures_tab import FixturesTab
+from .shows_tab import ShowsTab
+from .stage_tab import StageTab
+
+__all__ = ['BaseTab', 'ConfigurationTab', 'FixturesTab', 'ShowsTab', 'StageTab']
